+++ Revo.Core/Lifecycle/IApplicationLifecycleNotifier.cs
@@ -1,0 +1,9 @@
+﻿namespace Revo.Core.Lifecycle
+{
+    public interface IApplicationLifecycleNotifier
+    {
+        void NotifyStarting();
+        void NotifyStarted();
+        void NotifyStopping();
+    }
+}