+++ Revo.Core/Lifecycle/IApplicationStartListener.cs
@@ -1,7 +1,0 @@
-﻿namespace Revo.Core.Lifecycle
-{
-    public interface IApplicationStartListener
-    {
-        void OnApplicationStarted();
-    }
-}