+++ Revo.Core/Lifecycle/IApplicationStartListenerInitializer.cs
@@ -1,8 +1,0 @@
-﻿namespace Revo.Core.Lifecycle
-{
-    public interface IApplicationStartListenerInitializer
-    {
-        void InitializeStarted();
-        void DeinitializeStopping();
-    }
-}