+++ Revo.Core/Lifecycle/IApplicationStopListener.cs
@@ -1,7 +1,0 @@
-﻿namespace Revo.Core.Lifecycle
-{
-    public interface IApplicationStopListener
-    {
-        void OnApplicationStopping();
-    }
-}